requirements.txt
dialogflow==0.6.0
python-telegram-bot==11.1.0
vk_api==11.4.0